Web: Files: removed FormatsStore, added getting file formats from SettingsStore
