+++ products/ASC.Files/Client/src/pages/Home/Section/Body/TilesView/FileTile.js
@@ -108,8 +108,8 @@ const FilesTile = (props) => {
   );
 };
 
-export default inject(({ formatsStore }) => {
-  const { getIcon } = formatsStore.iconFormatsStore;
+export default inject(({ settingsStore }) => {
+  const { getIcon } = settingsStore;
   return { getIcon };
 })(
   withTranslation("Home")(
