--- products/ASC.Files/Client/src/store/FormatsStore.js
@@ -1,16 +1,0 @@
-import { makeAutoObservable } from "mobx";
-
-class FormatsStore {
-  iconFormatsStore;
-  mediaViewersFormatsStore;
-  docserviceStore;
-
-  constructor(iconFormatsStore, mediaViewersFormatsStore, docserviceStore) {
-    makeAutoObservable(this);
-    this.iconFormatsStore = iconFormatsStore;
-    this.mediaViewersFormatsStore = mediaViewersFormatsStore;
-    this.docserviceStore = docserviceStore;
-  }
-}
-
-export default FormatsStore;
