+++ products/ASC.Files/Client/src/store/MediaViewerDataStore.js
@@ -2,16 +2,16 @@ import { makeAutoObservable } from "mobx";
 
 class MediaViewerDataStore {
   filesStore;
-  formatsStore;
+  settingsStore;
 
   id = null;
   visible = false;
   previewFile = null;
 
-  constructor(filesStore, formatsStore) {
+  constructor(filesStore, settingsStore) {
     makeAutoObservable(this);
     this.filesStore = filesStore;
-    this.formatsStore = formatsStore;
+    this.settingsStore = settingsStore;
   }
 
   setMediaViewerData = (mediaData) => {
@@ -35,7 +35,7 @@ class MediaViewerDataStore {
   };
 
   get playlist() {
-    const { isMediaOrImage } = this.formatsStore.mediaViewersFormatsStore;
+    const { isMediaOrImage } = this.settingsStore;
     const { files } = this.filesStore;
 
     const filesList = [...files];
