+++ products/ASC.Files/Client/src/store/index.js
@@ -5,10 +5,6 @@ import TreeFoldersStore from "./TreeFoldersStore";
 import thirdPartyStore from "./ThirdPartyStore";
 import SettingsStore from "./SettingsStore";
 import FilesActionsStore from "./FilesActionsStore";
-import FormatsStore from "./FormatsStore";
-import iconFormatsStore from "./IconFormatsStore";
-import mediaViewersFormatsStore from "./MediaViewersFormatsStore";
-import docserviceStore from "./DocserviceStore";
 import MediaViewerDataStore from "./MediaViewerDataStore";
 import UploadDataStore from "./UploadDataStore";
 import SecondaryProgressDataStore from "./SecondaryProgressDataStore";
@@ -18,12 +14,6 @@ import VersionHistoryStore from "./VersionHistoryStore";
 import DialogsStore from "./DialogsStore";
 import selectedFilesStore from "./SelectedFilesStore";
 import store from "studio/store";
-
-const formatsStore = new FormatsStore(
-  iconFormatsStore,
-  mediaViewersFormatsStore,
-  docserviceStore
-);
 
 const selectedFolderStore = new SelectedFolderStore(store.auth.settingsStore);
 
@@ -38,11 +28,13 @@ const filesStore = new FilesStore(
   fileActionStore,
   selectedFolderStore,
   treeFoldersStore,
-  formatsStore,
   settingsStore,
   selectedFilesStore
 );
-const mediaViewerDataStore = new MediaViewerDataStore(filesStore, formatsStore);
+const mediaViewerDataStore = new MediaViewerDataStore(
+  filesStore,
+  settingsStore
+);
 
 const secondaryProgressDataStore = new SecondaryProgressDataStore();
 const primaryProgressDataStore = new PrimaryProgressDataStore();
@@ -54,7 +46,6 @@ const dialogsStore = new DialogsStore(
   selectedFolderStore
 );
 const uploadDataStore = new UploadDataStore(
-  formatsStore,
   treeFoldersStore,
   selectedFolderStore,
   filesStore,
@@ -82,7 +73,6 @@ const stores = {
   filesStore,
   settingsStore,
   mediaViewerDataStore,
-  formatsStore,
   versionHistoryStore,
   uploadDataStore,
   dialogsStore,
